backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
python-multipart
qrcode
pillow
python-dotenv
psycopg2-binary
asyncpg
